Doing client side formatting on property names that build out tables. Object tables are built by enumerating over properties of an object and presenting them with the property label as a key and it's value as a value:
JSON
{ FooBar: "Bar", Foo2: "Bar2" }

Object-table
--------------------
 Foo Bar  |  Bar
 Foo2     |  Bar2
-------------------

The formatting matches on any uppercase character preceded by a lower case character and inserts a space in-between them. Makes the interface much more readable and none of the processing needs to happen server side. PowerShell properties are passed through the web app unmodified and rendered nicely in the UI.